international/requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
